requirements.txt
biopython
hdbscan
matplotlib
numpy
pandas
plotly
scikit-learn
scipy
seaborn
torch
transformers
umap-learn
